+++ examples/bar.py
@@ -1,0 +1,5 @@
+from quo.console import Console
+
+console = Console()
+
+console.bar("ff")