+++ examples/full-screen/simple-demos/vertical-split.py
@@ -2,9 +2,9 @@
 """
 Vertical split example.
 """
-from quo.console import Console
+from quo import container
 from quo.keys import Bind
-from quo.layout import FormattedTextControl, Window, VSplit, Layout
+from quo.layout import FormattedTextControl, Window, VSplit
 from quo.text import Text
 
 # 1. The layout
@@ -15,17 +15,6 @@ right_text = Text("""
 
         <b>(right pane)</b>""")
 
-
-body = VSplit(
-    [
-        Window(FormattedTextControl(left_text)),
-        Window(width=1, char="|", style="fg:green"),  # Vertical line in the middle.
-        Window(FormattedTextControl(right_text)),
-    ]
-)
-
-
-# 2. Key bindings
 bind = Bind()
 
 @bind.add("q")
@@ -34,5 +23,13 @@ def _(event):
     event.app.exit()
 
 
-# 3. The `Application`
-Console(Layout(body), bind=bind, full_screen=True).run()
+container(
+        VSplit([
+            Window(FormattedTextControl(left_text)),
+            Window(width=1, char="|", style="fg:green"),  # Vertical line in the middle.
+            Window(FormattedTextControl(right_text))
+            ]),
+        bind=bind,
+        full=True)
+
+