+++ examples/progress/scrolling-task-name.py
@@ -4,11 +4,11 @@ A very simple progress bar where the name of the task scrolls, because it's too 
 iterator.
 """
 import time
-import quo
 
+from quo.progress import ProgressBar
 
 def main():
-    with quo.ProgressBar(
+    with ProgressBar(
             title="Scrolling task name (make sure the window is not too big).") as pb:
         for i in pb(range(800),
                 label="This is a very very very long task that requires horizontal scrolling ...",):