+++ examples/prompts/clock-input.py
@@ -18,7 +18,7 @@ def get_prompt():
             ("bg:cornsilk fg:maroon", "Enter something: ")
     ]
 
-session = Prompt(refresh_interval=0.5, style="style)"
+session = Prompt(refresh_interval=0.5)#, style="style")
 def main():
     result = session.prompt(get_prompt)
     echo(f"You said: {result}")