+++ src/quo/console/__init__.py
@@ -30,3 +30,5 @@ from quo.decorators import tether as tether
 #    "in_terminal",
 #    "run_in_terminal",
 #]
+
+console = Console()