+++ src/quo/keys/__init__.py
@@ -4,3 +4,5 @@ from quo.keys.focus import next, previous
 from quo.keys.key_binding.vi_state import InputMode
 
 KeyBinder = Bind
+
+bind = Bind()