+++ docs/changes.rst
@@ -17,6 +17,19 @@ Unreleased
 
 - :func:`quo.console.Console.pager`
 
+``Version 2022.3.4``
+---------------------
+
+Released on 2022-03-18
+
+**Added**
+^^^^^^^^^^^
+- Added :param:`bind` to :func:`quo.container`
+- Added :param:`focused_element` to :func:`quo.container`
+- Added :param:`full_screen` to :func:`quo.container`
+- Added :param:`mouse_support` to :func:`quo.container`
+- Added :param:`refresh` to :func:`quo.container`
+
 ``Version 2022.3.3``
 ---------------------